Update Task43.java
fixt int => double

+++ src/com/example/androidcourse/Task43.java
@@ -8,11 +8,11 @@ public class Task43 {
         System.out.println("Enter a: ");
         int a = in.nextInt();
         double sum = 0;
-        int i = 1;
+        double i = 1;
 
 
         while (i <= a) {
-            sum +=(double) (1 / i);
+            sum += 1 / i;
             i++;
         }
         System.out.println("SummaHarmon " + sum);
